Changed how opcodes are handled
Instead of one big long if..else thing, they are kept as symbols and a
list of map/values. This means that they can be precomputed, and we can
use a 64k lookup table (which takes a while to be computed initially, to
be fair...). It also allows reasoning with them, such as detecting when
two words map to the same opcode, which happens a lot, it seems.